Added haircuts

+++ db/migrate/20131001124053_create_transactions.rb
@@ -4,8 +4,8 @@ class CreateTransactions < ActiveRecord::Migration
     create_table :transactions do |t|
       t.belongs_to :appointment, null: false
       t.belongs_to :invoice, null: false
-      t.decimal :total, null: false
-      t.decimal :tip, null: false, default: 0
+      t.float :total, null: false
+      t.float :tip, null: false, default: 0
     end
   end
